backend/pyproject.toml
[project]
name = "fleetiq-backend"
version = "0.2.0"
description = "FleetIQ FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "pydantic-settings>=2.3.0",
  "psycopg[binary]>=3.1.19",
  "python-jose>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "email-validator>=2.2.0",
  "alembic>=1.13.2",
  "celery>=5.4.0",
  "redis>=5.0.0",
  "scikit-learn>=1.5.0",
  "numpy>=1.26.0",
  "joblib>=1.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
